Added large test fixtures to own file. Added test
